more response updates

+++ ServiceLayer/Service/ProductService.cs
@@ -1,8 +1,11 @@
 ﻿using AutoMapper;
 using DomainLayer.Models;
 using DomainLayer.Models.ViewModels;
+using Microsoft.AspNetCore.Mvc;
 using RepositoryLayer.IRepository;
+using ServiceLayer.Handlers;
 using ServiceLayer.IService;
+using System.Net;
 
 namespace ServiceLayer.Service
 {
@@ -17,33 +20,40 @@ namespace ServiceLayer.Service
             this.mapper = mapper;
         }
 
-        public void AddProduct(AddProductModel addProduct)
+        public Task<ObjectResult> AddProduct(AddProductModel addProduct)
         {
             var product = mapper.Map<Product>(addProduct);
             this.productRepository.AddProduct(product);
         }
 
-        public async Task<List<ProductsViewModel>> GetProducts()
+        public async Task<ObjectResult> GetProducts()
         {
             var products = await this.productRepository.GetAllAsync();
-            return mapper.Map<List<ProductsViewModel>>(products);
+            var response = new OkObjectResult(mapper.Map<List<ProductsViewModel>>(products));
+            return response;
         }
 
-        public async Task<ProductsViewModel> GetByIdAsync(Guid Id)
+        public async Task<ObjectResult> GetByIdAsync(Guid Id)
         {
             var product = await this.productRepository.GetByIdAsync(Id);
-            return mapper.Map<ProductsViewModel>(product);
+            var response = new ObjectResult(mapper.Map<ProductsViewModel>(product));
+            return response;
         }
 
-        public async Task<bool> DeleteAsync(Guid Id)
+        public async Task<ObjectResult> DeleteAsync(Guid Id)
         {
             var product = await this.productRepository.GetByIdAsync(Id);
+            if (product == null) 
+            {
+                throw new HttpStatusCodeException(HttpStatusCode.NotFound,
+                    "PRODUCT_NOT_FOUND");
+            }
 
             var result = await this.productRepository.DeleteAsync(product);
-            return result;
+            return result ? new ObjectResult("PRODUCT_HAS_BEEN_DELETED") : new ObjectResult("PRODUCT_HAS_NOT_BEEN_DELETED");
         }
 
-        public void UpdateProduct(UpdateProductModel updateProduct)
+        public Task<ObjectResult> UpdateProduct(UpdateProductModel updateProduct)
         {
             var productToUpdate = mapper.Map<Product>(updateProduct);
             this.productRepository.UpdateProduct(productToUpdate);
